Fixed the disappearing tile bug.
Some improvements for the Image States system, can now go from State A which is a child of state B (itself a child of R), to state C a child of R.

Flipping added, so now images can create two states, one for the "normal" view, and one for their X-Flipped counterpart. Especially useful for the difference in facing. (i.e. walking right being the flip of walking left).